Allow "any" as tray_output specifier
If no tray_output specified, i3bar automatically shows on the first
encountered display a tray_output.

But when specifying tray_output, it's not possible to add a fallback
tray_output, which matches any screen.

So specifying "tray_output primary" will require a primary monitor to
exist and be available. "any" provides an easy fallback.

Fixes #3317